% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/present_data.R
\name{contam_counts_figure}
\alias{contam_counts_figure}
\title{create contamination count figure}
\usage{
contam_counts_figure(db_con, genus)
}
\arguments{
\item{db_con}{peprDB connection}

\item{genus}{rm genus}
}
\description{
create contamination count figure
}

