% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/present_data.R
\name{contam_distribution_figure}
\alias{contam_distribution_figure}
\title{create contamination distribution figure}
\usage{
contam_distribution_figure(db_con, genus)
}
\arguments{
\item{db_con}{peprDB connection}

\item{genus}{rm genus}
}
\description{
create contamination distribution figure
}

