% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/present_contamination.R
\name{contam_heatmap_figure}
\alias{contam_heatmap_figure}
\title{create contaminant heatmap}
\usage{
contam_heatmap_figure(db_con, genus)
}
\arguments{
\item{db_con}{peprDB connection}

\item{genus}{rm genus}
}
\description{
create contaminant heatmap
}

