% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_data.R
\name{coverage_table}
\alias{coverage_table}
\title{Per sample coverage}
\usage{
coverage_table(db_con, platforms = c("miseq", "pgm"))
}
\arguments{
\item{db_con}{}

\item{platforms}{}
}
\value{
adds coverage table to db
}
\description{
Per sample coverage
}
\examples{
coverage_table(peprDB)
}

