% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/present_homogeneity.R
\name{create_pairwise_homogeneity_df}
\alias{create_pairwise_homogeneity_df}
\title{data frame for pairwise homogeneity snp results}
\usage{
create_pairwise_homogeneity_df(db_con)
}
\arguments{
\item{db_con}{peprDB connection}
}
\value{
data_frame
}
\description{
data frame for pairwise homogeneity snp results
}

