% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_data.R
\name{createPeprDB}
\alias{createPeprDB}
\title{Generates sqlite database from pipeline output}
\usage{
createPeprDB(db_path, param_yaml, qc_stats_dir, homogeneity_dir, consensus_dir,
  purity_dir, pilon_dir)
}
\arguments{
\item{db_path}{path for database}

\item{qc_stats_dir}{directory with qc metrics}

\item{homogeneity_dir}{directory with homogeneity results}

\item{consensus_dir}{directory with consensus results}

\item{purity_dir}{directory with purity results}

\item{pilon_dir}{directory with pilon results}

\item{params_yaml}{yaml pipeline configuration file}
}
\description{
Generates sqlite database from pipeline output
}

