% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/present_homogeneity.R
\name{homogeneity_freq_figure}
\alias{homogeneity_freq_figure}
\title{Figure with distribution of variant frequency for pair-wise variants}
\usage{
homogeneity_freq_figure(db_con, prop_pairs_cutoff = 0.25, freq_cutoff = 99)
}
\arguments{
\item{db_con}{peprDB connection}

\item{prop_pairs_cutoff}{cutoff values for positions with proportions of pairs with variant call default 0.25}

\item{freq_cutoff}{cutoff value for positions variant frequency default 0.98}
}
\value{
figure
}
\description{
Figure with distribution of variant frequency for pair-wise variants
}

