% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/present_homogeneity.R
\name{homogeneity_sig_table}
\alias{homogeneity_sig_table}
\title{data frame for pairwise homogeneity snp results}
\usage{
homogeneity_sig_table(db_con, rename_cols = FALSE)
}
\arguments{
\item{db_con}{peprDB connection}

\item{rename_col}{boolean if TRUE rename columns for use in ROA table}
}
\value{
data_frame
}
\description{
data frame for pairwise homogeneity snp results
}

