% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_data.R
\name{init_peprDB}
\alias{init_peprDB}
\title{Initiates sqlite database for storing pepr pipeline output data.}
\usage{
init_peprDB(data_dir = NULL, db_path = NULL, create = TRUE)
}
\arguments{
\item{data_dir}{data directory generated by the pepr pipeline,
will generate new database in directory if db_path is not provided.}

\item{db_path}{optional path to initiate the database}

\item{Set}{to FALSE to have database be read-only, default TRUE}
}
\value{
dplyr database ....
}
\description{
Initiates sqlite database for storing pepr pipeline output data.
}
\examples{
init_peprDB("/path/to/PEPR-DATA")
init_peprDB(db_path = "/path/to/peprDB.sqlite", create = TRUE)
}

