% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_fastqc}
\alias{load_fastqc}
\title{Loads sequence dataset metric files generated by fastqc into database}
\usage{
load_fastqc(metrics_dir, db_con)
}
\arguments{
\item{metrics_dir}{directory with fastqc output}

\item{db_con}{database connection}
}
\description{
Loads sequence dataset metric files generated by fastqc into database
}
\examples{
load_fastqc("/path/to/metrics_dir", peprDB)
}

