% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_metrics}
\alias{load_metrics}
\title{Loads sequence dataset metric file generated by picard into database}
\usage{
load_metrics(metrics_dir, db_con)
}
\arguments{
\item{metrics_dir}{directory with picard metrics output}

\item{db_con}{database connection}
}
\description{
Loads sequence dataset metric file generated by picard into database
}
\examples{
load_metrics("/path/to/metrics_dir", peprDB)
}

