% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_peprMeta}
\alias{load_peprMeta}
\title{Loads metadata file from pepr pipeline into database}
\usage{
load_peprMeta(metadata, db_con)
}
\arguments{
\item{metadata}{yaml file with pipeline metadata}

\item{db_con}{optional path to initiate the database}
}
\value{
database
}
\description{
Loads metadata file from pepr pipeline into database
}
\examples{
load_peprMeta("/path/to/metadata.yaml", pepr_con)
}

