% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_pilon}
\alias{load_pilon}
\title{Loads pilon output data into sqlite db}
\usage{
load_pilon(pilon_dir, db_con)
}
\arguments{
\item{pilon_dir}{directory with pilon results}

\item{db_con}{database connection}
}
\description{
Loads pilon output data into sqlite db
}
\examples{
load_pilon("/path/to/purity_dir", peprDB)
}

