% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_purity}
\alias{load_purity}
\title{Loads pathoscope output data into sqlite db}
\usage{
load_purity(purity_dir, db_con)
}
\arguments{
\item{purity_dir}{directory with pathoscope output}

\item{db_con}{database connection}
}
\description{
Loads pathoscope output data into sqlite db
}
\examples{
load_purity("/path/to/purity_dir", peprDB)
}

