% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_varscan}
\alias{load_varscan}
\title{Loads varscan output data into sqlite db}
\usage{
load_varscan(homogeneity_dir, db_con)
}
\arguments{
\item{homogeneity_dir}{directory with varscan results}

\item{db_con}{database connection}
}
\description{
Loads varscan output data into sqlite db
}
\examples{
load_pilon("/path/to/homogeneity_dir", peprDB)
}

