% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/present_data.R
\name{seq_summary_table}
\alias{seq_summary_table}
\title{create df for dataset summary table}
\usage{
seq_summary_table(db_con)
}
\arguments{
\item{db_con}{peprDB connection}
}
\description{
create df for dataset summary table
}

