% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_consensus}
\alias{load_consensus}
\title{Loads full genome vcf output data into sqlite db}
\usage{
load_consensus(consensus_dir, db_con)
}
\arguments{
\item{consensus_dir}{directory with consensus results}

\item{db_con}{database connection}
}
\description{
Loads full genome vcf output data into sqlite db
}
\examples{
load_pilon("/path/to/purity_dir", peprDB)
}

